requirements.txt
streamlit
python-dotenv
langchain-core
langchain-openai
pandas
requests
typing